fix: 94 - evaluate running tests with postgres backend
* not complete, could be a significant amount of work.
* thinking leave things in place as unit testing
* postgres testing would become part of integration testing

+++ server/backend/tests/conftest.py
@@ -13,7 +13,7 @@ https://fastapi.tiangolo.com/advanced/testing-database/
 import os
 import sys
 
-sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
 # sys.path.append(os.path.dirname(__file__))
 
 import logging
@@ -31,10 +31,13 @@ from fastapi.testclient import TestClient
 from sqlalchemy import create_engine, event
 from sqlalchemy.engine.base import Engine
 from sqlalchemy.orm import sessionmaker
+import dotenv
 
 # global placeholder to be populated by fixtures for database test
 # sessions, required to override the get_db method.
 testSession = None
+
+# debug:
 
 LOGGER = logging.getLogger(__name__)
 
@@ -46,12 +49,21 @@ pytest_plugins = [
     "fixtures.fixtures_crud_user",
     "fixtures.fixtures_router_user",
     "fixtures.fixtures_crud_role",
-    "fixtures.fixtures_crud_user_role_assignment"
+    "fixtures.fixtures_crud_user_role_assignment",
 ]
 
 
+@pytest.fixture(scope="session")
+def loadEnvs():
+    #'server/backend/tests/env-testing'
+    testingDotEnvFilePath = os.path.join(
+        os.path.dirname(__file__), 'env-testing')
+    dotenv.load_dotenv(testingDotEnvFilePath)
+    usePostgres = os.getenv("USE_POSTGRES", "false")
+    LOGGER.debug(f"env: {usePostgres}")
+
 @pytest.fixture(scope="function")
-def getApp(sessionObjects, dbEngine: Engine) -> Generator[FastAPI, Any, None]:
+def getApp(sessionObjects, loadEnvs, dbEngine: Engine) -> Generator[FastAPI, Any, None]:
     """
     Create a fresh database on each test case.
     """
@@ -62,14 +74,17 @@ def getApp(sessionObjects, dbEngine: Engine) -> Generator[FastAPI, Any, None]:
     yield app
 
 
-# This @event is important. By default FOREIGN KEY constraints have no effect on the operation of the table from SQLite.
+# This @event is important. By default FOREIGN KEY constraints have no effect on
+# the operation of the table from SQLite.
 # It (FOREIGN KEY) only works when emitting CREATE statements for tables.
-# Reference: https://docs.sqlalchemy.org/en/14/dialects/sqlite.html#foreign-key-support
+# Reference:
+#   https://docs.sqlalchemy.org/en/14/dialects/sqlite.html#foreign-key-support
 @event.listens_for(Engine, "connect")
 def set_sqlite_pragma(dbapi_connection, connection_record):
-    cursor = dbapi_connection.cursor()
-    cursor.execute("PRAGMA foreign_keys=ON")
-    cursor.close()
+    if os.getenv("USE_POSTGRES", "false") == "false":
+        cursor = dbapi_connection.cursor()
+        cursor.execute("PRAGMA foreign_keys=ON")
+        cursor.close()
 
 
 @pytest.fixture(scope="function")
@@ -95,7 +110,7 @@ def sessionObjects(dbEngine: Engine) -> sessionmaker:
 
 
 @pytest.fixture(scope="module")
-def dbEngine() -> Engine:
+def dbEngine(loadEnvs) -> Engine:
     # should re-create the database every time the tests are run, the following
     # line ensure database that maybe hanging around as a result of a failed
     # test is deleted
@@ -103,11 +118,10 @@ def dbEngine() -> Engine:
         LOGGER.debug("remove the database: ./test_db.db'")
         os.remove("./test_db.db")
 
-    execution_options = {"schema_translate_map": {"app_fam": None}}
 
+    if os.getenv("USE_POSTGRES", "false") == "false":
+        execution_options = {"schema_translate_map": {"app_fam": None}}
 
-
-    if os.getenv("USE_POSTGRES", "false") == 'false':
         SQLALCHEMY_DATABASE_URL = "sqlite:///./test_db.db"
         LOGGER.debug(f"SQL Alchemy URL: {SQLALCHEMY_DATABASE_URL}")
         engine = create_engine(
@@ -115,26 +129,22 @@ def dbEngine() -> Engine:
             connect_args={"check_same_thread": False},
             execution_options=execution_options,
         )
+        model.Base.metadata.create_all(bind=engine)
     else:
         SQLALCHEMY_DATABASE_URL = config.getDBString()
-        engine = create_engine(
-            SQLALCHEMY_DATABASE_URL,
-            echo=False)
+        engine = create_engine(SQLALCHEMY_DATABASE_URL, echo=False)
+        # assumes migrations have been run
 
-
-
-    model.Base.metadata.create_all(bind=engine)
     LOGGER.debug(f"engine type: {type(engine)}")
     yield engine
 
-    # dropping all objects in the test database and...
-    # delete the test database
-
-    model.Base.metadata.drop_all(engine)
-    if os.path.exists("./test_db.db"):
-        LOGGER.debug("remove the database: ./test_db.db'")
-        os.remove("./test_db.db")
-
+    if os.getenv("USE_POSTGRES", "false") == "false":
+        # dropping all objects in the test database and...
+        # delete the test database
+        model.Base.metadata.drop_all(engine)
+        if os.path.exists("./test_db.db"):
+            LOGGER.debug("remove the database: ./test_db.db'")
+            os.remove("./test_db.db")
 
 
 @pytest.fixture(scope="function")
@@ -170,14 +180,16 @@ def override_get_db():
 
 def getFixtureParams(request):
     """
-    Helper function to pass custom param into fixture to do setup/dear-down logic mostly.
-    For example, in test, can mark the test that uses some fixture to pass param into the fixture:
+    Helper function to pass custom param into fixture to do setup/dear-down
+    logic mostly.  For example, in test, can mark the test that uses some
+    fixture to pass param into the fixture:
         @pytest.mark.fixture_data({"clean_up": False})
 
     And then in that specific fixture, get the param individually by calling:
         need_cleanup = getFixtureParams(request)['clean_up']
 
-        Note, the fixture needs to have pytest 'request' as the argument passing into the function.
+        Note, the fixture needs to have pytest 'request' as the argument passing
+        into the function.
     """
     marker = request.node.get_closest_marker("fixture_data")
     params = marker.args[0]
